handle refresh gnosisjs connection
