+++ src/utils/helpers.js
@@ -3,6 +3,8 @@
 import { mapValues, startsWith, isArray } from 'lodash'
 import Decimal from 'decimal.js'
 import { HEX_VALUE_REGEX, OUTCOME_TYPES } from 'utils/constants'
+import { WALLET_PROVIDER } from 'integrations/constants'
+import Web3 from 'web3'
 
 export const hexWithoutPrefix = (value) => {
   if (HEX_VALUE_REGEX.test(value)) {
@@ -143,3 +145,20 @@ export const isModerator = accountAddress => (
 )
 
 export const getModerators = () => process.env.WHITELIST
+
+export const getGnosisJsOptions = (provider) => {
+  const opts = {}
+
+  if (provider && provider.name === WALLET_PROVIDER.METAMASK) {
+    // Inject window.web3
+    opts.ethereum = window.web3.currentProvider
+  } else if (provider && provider === WALLET_PROVIDER.PARITY) {
+    // Inject window.web3
+    opts.ethereum = window.web3.currentProvider
+  } else {
+    // Default remote node
+    opts.ethereum = new Web3(new Web3.providers.HttpProvider(`${process.env.ETHEREUM_URL}`)).currentProvider
+  }
+
+  return opts
+}
